rename all the other dodgy named files too
